requirements.txt
chempy==0.9.0
Django==5.1
djangorestframework==3.15.2
drf_spectacular==0.27.2
python-dotenv==1.0.1
rdkit==2024.3.3
Requests==2.32.3
celery==5.4.0
msgpack==1.0.8
amqp==5.2.0
django-cors-headers==4.5.0
django-celery-results==2.5.1
django-filter==24.3
psycopg==3.2.3
psycopg-binary==3.2.3
pymemcache==4.0.0
django-rest-knox==5.0.2
email-validator==2.2.0
bcrypt==4.2.0
pyfsig==1.0.0
django-clamd==1.0.0